requirements.txt
python-dotenv==1.1.0
selenium==4.31.0
tweepy==4.15.0
